Move code and fix 'fragile base class' bug in pydantic model
