+++ cumulusci/core/config/project_config.py
@@ -37,7 +37,6 @@ from cumulusci.core.github import (
     get_github_api_for_repo,
 )
 from cumulusci.core.source import GitHubSource, LocalFolderSource, NullSource
-from cumulusci.core.utils import merge_cumulus_config
 from cumulusci.utils.fileutils import FSResource, open_fs_resource
 from cumulusci.utils.git import current_branch, git_path, parse_repo_url, split_repo_url
 from cumulusci.utils.yaml.cumulusci_yml import (
@@ -45,6 +44,7 @@ from cumulusci.utils.yaml.cumulusci_yml import (
     LocalFolderSourceModel,
     cci_safe_load,
 )
+from cumulusci.utils.yaml.merge_cumulus_yaml import merge_cumulus_config
 
 if TYPE_CHECKING:
     from cumulusci.core.config.universal_config import UniversalConfig
