+++ cumulusci/core/config/universal_config.py
@@ -6,8 +6,8 @@ from cumulusci.core.config.project_config import (
     BaseProjectConfig,
     ProjectConfigPropertiesMixin,
 )
-from cumulusci.core.utils import merge_cumulus_config
 from cumulusci.utils.yaml.cumulusci_yml import cci_safe_load
+from cumulusci.utils.yaml.merge_cumulus_yaml import merge_cumulus_config
 
 __location__ = os.path.dirname(os.path.realpath(__file__))
