+++ cumulusci/core/tests/test_utils_merge_config.py
@@ -4,6 +4,7 @@ import pytest
 
 from cumulusci.core import utils
 from cumulusci.core.exceptions import ConfigMergeError, CumulusCIException
+from cumulusci.utils.yaml.merge_cumulus_yaml import merge_cumulus_config
 
 
 @pytest.fixture
@@ -33,7 +34,7 @@ def universal_config():
 
 
 def test_init():
-    config = utils.merge_cumulus_config(
+    config = merge_cumulus_config(
         {
             "universal_config": {"hello": "world"},
             "user_config": {"hello": "christian"},
@@ -44,7 +45,7 @@ def test_init():
 
 def test_merge_failure():
     with pytest.raises(ConfigMergeError) as e:
-        utils.merge_cumulus_config(
+        merge_cumulus_config(
             {
                 "universal_config": {"hello": "world", "test": {"sample": 1}},
                 "user_config": {"hello": "christian", "test": [1, 2]},
