+++ cumulusci/utils/yaml/cumulusci_yml.py
@@ -61,7 +61,7 @@ class Step(CCIDictModel):
 
 class Task(CCIDictModel):
     class_path: str = None
-    extends: str = None
+    extends: Optional[str] = None
     description: str = None
     group: str = None
     # additionalProperties here works around an
@@ -71,9 +71,10 @@ class Task(CCIDictModel):
     name: str = None  # get rid of this???
 
     @root_validator
-    def validate(cls, values: dict):
+    def validate_exclusive_keys(cls, values: dict):
+        print(values)
         assert not (
-            "extends" in values and "class_path" in values
+            values.get("extends") and values.get("class_path")
         ), "Please do not include both `class_path` and `extends` for the same task"
 
         return values
@@ -322,7 +323,7 @@ def cci_safe_load(
     logger = logger or default_logger
 
     with load_from_source(source) as (data_stream, filename):
-        data: dict = load_yaml_data(data_stream, filename)
+        data = load_yaml_data(data_stream, filename)
         context = context or filename
 
         try:
